refactor(model)!: move related modules under guild
* move invite, scheduled_event, template -> guild
* hide `MemberIntermediary` (like other `*Intermediary` types)
* use `crate::guild` over `super::super` in `AuditLogEventType` doc
comment path

Part of #1791
