+++ twilight-http/src/request/guild/update_guild_welcome_screen.rs
@@ -7,8 +7,8 @@ use crate::{
 };
 use serde::Serialize;
 use twilight_model::{
+    guild::invite::{WelcomeScreen, WelcomeScreenChannel},
     id::{marker::GuildMarker, Id},
-    invite::{WelcomeScreen, WelcomeScreenChannel},
 };
 
 #[derive(Serialize)]
