+++ twilight-http/src/response/mod.rs
@@ -650,10 +650,7 @@ impl Future for MemberFuture {
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         match Pin::new(&mut self.future).poll(cx) {
             Poll::Ready(Ok(mut bytes)) => {
-                let mut deserializer = match json_deserializer(&mut bytes) {
-                    Ok(deserializer) => deserializer,
-                    Err(source) => return Poll::Ready(Err(source)),
-                };
+                let mut deserializer = json_deserializer(&mut bytes)?;
                 let member_deserializer = MemberDeserializer::new(self.guild_id);
 
                 let result = member_deserializer
@@ -725,10 +722,7 @@ impl Future for MemberListFuture {
     fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         match Pin::new(&mut self.0.future).poll(cx) {
             Poll::Ready(Ok(mut bytes)) => {
-                let mut deserializer = match json_deserializer(&mut bytes) {
-                    Ok(deserializer) => deserializer,
-                    Err(source) => return Poll::Ready(Err(source)),
-                };
+                let mut deserializer = json_deserializer(&mut bytes)?;
                 let member_list_deserializer = MemberListDeserializer::new(self.0.guild_id);
 
                 let result = member_list_deserializer
@@ -887,7 +881,7 @@ mod tests {
     async fn test_decompression() -> Result<(), Box<dyn Error + Send + Sync>> {
         use super::decompress;
         use hyper::Body;
-        use twilight_model::invite::Invite;
+        use twilight_model::guild::invite::Invite;
 
         const COMPRESSED: [u8; 685] = [
             3, 160, 2, 0, 228, 178, 169, 189, 190, 59, 251, 86, 18, 36, 232, 63, 98, 235, 98, 82,
