+++ twilight-model/src/channel/thread/mod.rs
@@ -1,11 +1,9 @@
 mod auto_archive_duration;
 mod listing;
-mod member;
+pub(crate) mod member;
 mod metadata;
 
 pub use self::{
     auto_archive_duration::AutoArchiveDuration, listing::ThreadsListing, member::ThreadMember,
     metadata::ThreadMetadata,
 };
-
-pub(crate) use self::member::ThreadMemberIntermediary;
