+++ twilight-model/src/gateway/payload/incoming/guild_scheduled_event_delete.rs
@@ -1,10 +1,10 @@
-use crate::scheduled_event::GuildScheduledEvent;
+use crate::guild::scheduled_event::GuildScheduledEvent;
 use serde::{Deserialize, Serialize};
 use std::ops::{Deref, DerefMut};
 
 /// Sent when a guild scheduled event is deleted. The inner payload is a [`GuildScheduledEvent`].
 ///
-/// [`GuildScheduledEvent`]: crate::scheduled_event::GuildScheduledEvent
+/// [`GuildScheduledEvent`]: crate::guild::scheduled_event::GuildScheduledEvent
 #[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
 pub struct GuildScheduledEventDelete(pub GuildScheduledEvent);
