+++ twilight-model/src/gateway/payload/incoming/thread_members_update.rs
@@ -1,5 +1,5 @@
 use crate::{
-    channel::thread::{ThreadMember, ThreadMemberIntermediary},
+    channel::thread::member::{ThreadMember, ThreadMemberIntermediary},
     id::{
         marker::{ChannelMarker, GuildMarker, UserMarker},
         Id,
