+++ twilight-model/src/guild/audit_log/event_type.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 pub enum AuditLogEventType {
     /// [Guild] was updated.
     ///
-    /// [Guild]: super::super::Guild
+    /// [Guild]: crate::guild::Guild
     GuildUpdate,
     /// [Channel] was created.
     ///
@@ -39,56 +39,56 @@ pub enum AuditLogEventType {
     ChannelOverwriteDelete,
     /// [Member] was kicked.
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     MemberKick,
     /// [Member] prune began.
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     MemberPrune,
     /// [Member] was banned.
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     MemberBanAdd,
     /// [Member]'s [ban] was removed.
     ///
-    /// [ban]: super::super::Ban
-    /// [Member]: super::super::Member
+    /// [ban]: crate::guild::Ban
+    /// [Member]: crate::guild::Member
     MemberBanRemove,
     /// [Member] was updated.
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     MemberUpdate,
     /// [Member] either had a [role] attached or removed.
     ///
-    /// [Member]: super::super::Member
-    /// [role]: super::super::Role
+    /// [Member]: crate::guild::Member
+    /// [role]: crate::guild::Role
     MemberRoleUpdate,
     /// [Member] was moved between voice [channel]s.
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     /// [channel]: crate::channel::Channel
     MemberMove,
     /// [Member] was disconnected from a voice [channel].
     ///
-    /// [Member]: super::super::Member
+    /// [Member]: crate::guild::Member
     /// [channel]: crate::channel::Channel
     MemberDisconnect,
     /// [Bot user] was added to a [guild].
     ///
     /// [Bot user]: crate::user::User::bot
-    /// [guild]: super::super::Guild
+    /// [guild]: crate::guild::Guild
     BotAdd,
     /// [Role] was created.
     ///
-    /// [Role]: super::super::Role
+    /// [Role]: crate::guild::Role
     RoleCreate,
     /// [Role] was updated.
     ///
-    /// [Role]: super::super::Role
+    /// [Role]: crate::guild::Role
     RoleUpdate,
     /// [Role] was deleted.
     ///
-    /// [Role]: super::super::Role
+    /// [Role]: crate::guild::Role
     RoleDelete,
     /// [Invite] was created.
     ///
@@ -116,15 +116,15 @@ pub enum AuditLogEventType {
     WebhookDelete,
     /// [Emoji] was created.
     ///
-    /// [Emoji]: super::super::Emoji
+    /// [Emoji]: crate::guild::Emoji
     EmojiCreate,
     /// [Emoji] was updated.
     ///
-    /// [Emoji]: super::super::Emoji
+    /// [Emoji]: scrate::guild::Emoji
     EmojiUpdate,
     /// [Emoji] was deleted.
     ///
-    /// [Emoji]: super::super::Emoji
+    /// [Emoji]: crate::guild::Emoji
     EmojiDelete,
     /// [Message] was deleted.
     ///
@@ -146,15 +146,15 @@ pub enum AuditLogEventType {
     MessageUnpin,
     /// [Integration] was created.
     ///
-    /// [Integration]: super::super::GuildIntegration
+    /// [Integration]: crate::guild::GuildIntegration
     IntegrationCreate,
     /// [Integration] was updated.
     ///
-    /// [Integration]: super::super::GuildIntegration
+    /// [Integration]: crate::guild::GuildIntegration
     IntegrationUpdate,
     /// [Integration] was deleted.
     ///
-    /// [Integration]: super::super::GuildIntegration
+    /// [Integration]: crate::guild::GuildIntegration
     IntegrationDelete,
     /// [Stage instance] was created.
     ///
@@ -182,15 +182,15 @@ pub enum AuditLogEventType {
     StickerDelete,
     /// [`GuildScheduledEvent`] was created.
     ///
-    /// [`GuildScheduledEvent`]: crate::scheduled_event::GuildScheduledEvent
+    /// [`GuildScheduledEvent`]: crate::guild::GuildScheduledEvent
     GuildScheduledEventCreate,
     /// [`GuildScheduledEvent`] was updated.
     ///
-    /// [`GuildScheduledEvent`]: crate::scheduled_event::GuildScheduledEvent
+    /// [`GuildScheduledEvent`]: crate::guild::GuildScheduledEvent
     GuildScheduledEventUpdate,
     /// [`GuildScheduledEvent`] was deleted.
     ///
-    /// [`GuildScheduledEvent`]: crate::scheduled_event::GuildScheduledEvent
+    /// [`GuildScheduledEvent`]: crate::guild::GuildScheduledEvent
     GuildScheduledEventDelete,
     /// Thread [channel] was created.
     ///
