+++ twilight-model/src/guild/mod.rs
@@ -1,6 +1,19 @@
+//! Group of [`Channel`]s and [`User`]s with additional moderation options.
+//!
+//! [`Guild`]s allow, for example, assigning [`Role`]s to [`Member`]s to limit
+//! their [`Permissions`] globally, or per [`Channel`].
+//!
+//! Advanced user's may look inside the [`member`] module for custom
+//! [`Member`] deserializers.
+//!
+//! [`User`]: super::user::User
+
 pub mod audit_log;
 pub mod auto_moderation;
+pub mod invite;
 pub mod member;
+pub mod scheduled_event;
+pub mod template;
 
 mod ban;
 mod default_message_notification_level;
@@ -28,12 +41,16 @@ mod vanity_url;
 mod verification_level;
 mod widget;
 
+// `Member` should appear inline, as the `member` module is only for advanced
+// use. Public documentation is not available for re-exports.
+#[doc(inline)]
+pub use self::member::Member;
 pub use self::{
     ban::Ban, default_message_notification_level::DefaultMessageNotificationLevel, emoji::Emoji,
     explicit_content_filter::ExplicitContentFilter, feature::GuildFeature, info::GuildInfo,
     integration::GuildIntegration, integration_account::IntegrationAccount,
     integration_application::IntegrationApplication,
-    integration_expire_behavior::IntegrationExpireBehavior, member::Member, mfa_level::MfaLevel,
+    integration_expire_behavior::IntegrationExpireBehavior, mfa_level::MfaLevel,
     nsfw_level::NSFWLevel, partial_guild::PartialGuild, partial_member::PartialMember,
     permissions::Permissions, premium_tier::PremiumTier, preview::GuildPreview, prune::GuildPrune,
     role::Role, role_tags::RoleTags, system_channel_flags::SystemChannelFlags,
