+++ twilight-model/src/guild/scheduled_event/user.rs
