+++ twilight-model/src/guild/template/guild.rs
