+++ twilight-model/src/guild/template/mod.rs
