+++ twilight-model/src/guild/template/role.rs
