+++ twilight-model/src/lib.rs
@@ -23,10 +23,7 @@ pub mod gateway;
 pub mod guild;
 pub mod http;
 pub mod id;
-pub mod invite;
 pub mod oauth;
-pub mod scheduled_event;
-pub mod template;
 pub mod user;
 pub mod util;
 pub mod voice;
